SSE Util fixes for Clang
When compiling the SSE util code with clang, we would get an
error message that the mode parameter of the SSE4_cmpestrm
method cannot be used as an immediate value. The reason for this
is that Clang performs the analysis for the inline assembly before
it runs the inline pass. Even if the mode parameter is only used with
compile time constant values it will fail compilation.

The fix is to make this method a template method.